Update TNG build system
GROMACS can now find and use a copy of TNG installed on the system.
The default is still to compile the version shipped with GROMACS.
That compilation is now handled using the BuildTNG.cmake script, so
there is no duplication between the GROMACS and TNG CMake code that
builds TNG, and the result is fully portable with all supported
environments and CMake versions.

Moved TNG handling into src/gromacs/CMakeLists.txt where it is
actually used.

Bumped the version of the TNG library to 1.6. Corresponds to commit
932c30a0da55a581371cfce1453e5557a434bf42 in the TNG repository.

Renamed md5 things with GROMACS prefixes, so that there is no internal
name clash with the md5 things in TNG, and no client of libgromacs can
get unexpected behaviour.

Refs #1453 (partial fix)
Fixes #1520

+++ src/gromacs/fileio/tngio_for_tools.cpp
@@ -44,7 +44,7 @@
 #include "trx.h"
 
 #ifdef GMX_USE_TNG
-#include "../../external/tng_io/include/tng_io.h"
+#include "tng/tng_io.h"
 #endif
 
 #include "gromacs/legacyheaders/types/atoms.h"
@@ -427,7 +427,7 @@ gmx_bool gmx_read_next_tng_frame(tng_trajectory_t            input,
     void                   *values        = NULL;
     double                  frameTime     = -1.0;
     int                     size, blockDependency;
-    float                   prec;
+    double                  prec;
     const int               defaultNumIds = 5;
     static gmx_int64_t      fallbackRequestedIds[defaultNumIds] =
     {
@@ -778,7 +778,7 @@ gmx_bool gmx_get_tng_data_next_frame_of_block_type(tng_trajectory_t     input,
     gmx_int64_t         codecId;
     int                 blockDependency;
     void               *data = 0;
-    float               localPrec;
+    double              localPrec;
 
     stat = tng_data_block_name_get(input, blockId, name, maxLen);
     if (stat != TNG_SUCCESS)
